Moved the button to the Footer component so it is available in most pages, and included a fadeIn effect on the button when scrolling
