+++ src/components/Footer.js
@@ -1,10 +1,31 @@
-import React from "react";
+import React, { useEffect } from "react";
 import styles from "../styles/footer.module.scss";
 
 import {FaJira} from "react-icons/fa";
 import meh from "../assets/meh.png";
+import { FiChevronUp } from "react-icons/fi";
 
 export default function Footer(){
+
+  useEffect(() => {
+    window.onscroll = function(){scrollFunction()};
+  }, []);
+
+  const scrollFunction = () => {
+    const btnTop = document.getElementById('btnTop');
+    
+    if(document.body.scrollTop > 20 || document.documentElement.scrollTop > 20){
+      btnTop.style.display = "flex";
+    } else {
+      btnTop.style.display = "none";
+    }
+  }
+
+  const topFunction = () => {
+    document.body.scrollTop = 0;
+    document.documentElement.scrollTop = 0;
+  }
+
     return (
       <div className={styles.footer}>
         <a
@@ -33,6 +54,9 @@ export default function Footer(){
           winstall is not associated with Microsoft, Windows, or Windows Package
           Manager.
         </p>
+        <span id="btnTop" title="Go to top" onClick={topFunction} className={styles.backToTop}>
+          <FiChevronUp />
+        </span>
       </div>
     );
 }
