+++ src/styles/footer.module.scss
@@ -29,3 +29,64 @@
     opacity: 0.75;
     font-size: 18px;
 }
+
+.backToTop {
+    display: none;
+    position: fixed;
+    bottom: 10px;
+    right: 20px;
+    z-index: 99;
+    border: none;
+    outline: none;
+    background-color: var(--card-bg);
+    color: var(--textColor);
+    cursor: pointer;
+    border-radius:50%;
+    transition: 0.3s all ease-in-out;
+    box-shadow: 0px 3px 5px rgba(0, 0, 0, 0.25);
+    opacity: 0.75;
+    text-align: center;
+    width: 50px;
+    height: 50px;
+    justify-content: center;
+    font-size: 22px;
+    animation: fadeIn ease 1s;
+    -webkit-animation: fadeIn ease 1s;
+    -moz-animation: fadeIn ease 1s;
+    -o-animation: fadeIn ease 1s;
+    -ms-animation: fadeIn ease 1s;
+
+    svg{
+        padding-top: 15px;
+    }
+
+    &:hover, &:focus{
+        opacity: 1;
+        color: var(--textColor) !important;
+    }
+}
+
+@keyframes fadeIn {
+    0% {opacity:0;}
+    100% {opacity:0.75;}
+}
+
+@-moz-keyframes fadeIn {
+    0% {opacity:0;}
+    100% {opacity:0.75;}
+}
+    
+@-webkit-keyframes fadeIn {
+    0% {opacity:0;}
+    100% {opacity:0.75;}
+}
+    
+@-o-keyframes fadeIn {
+    0% {opacity:0;}
+    100% {opacity:0.75;}
+}
+    
+@-ms-keyframes fadeIn {
+    0% {opacity:0;}
+    100% {opacity:0.75;}
+}
