+++ src/styles/store.module.scss
@@ -101,34 +101,3 @@ ul{
         padding: 10px;
     }
 }
-
-.backToTop {
-    display: none;
-    position: fixed;
-    bottom: 20px;
-    right: 30px;
-    z-index: 99;
-    border: none;
-    outline: none;
-    background-color: var(--card-bg);
-    color: var(--textColor);
-    cursor: pointer;
-    border-radius:50%;
-    transition: 0.3s all ease-in-out;
-    box-shadow: 0px 3px 5px rgba(0, 0, 0, 0.25);
-    opacity: 0.75;
-    text-align: center;
-    width: 50px;
-    height: 50px;
-    justify-content: center;
-    font-size: 22px;
-
-    svg{
-        padding-top: 15px;
-    }
-
-    &:hover, &:focus{
-        opacity: 1;
-        color: var(--textColor) !important;
-    }
-}
